requirements.txt
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
et-xmlfile==1.0.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
pdfminer.six==20201018
pycparser==2.20
pycryptodome==3.10.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sortedcontainers==2.3.0
XlsxWriter==1.3.8
